working on main.ts from work

+++ frontend/src/components/earnings.js
@@ -1,4 +1,27 @@
 "use strict";
+var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
+    if (k2 === undefined) k2 = k;
+    var desc = Object.getOwnPropertyDescriptor(m, k);
+    if (!desc || ("get" in desc ? !m.__esModule : desc.writable || desc.configurable)) {
+      desc = { enumerable: true, get: function() { return m[k]; } };
+    }
+    Object.defineProperty(o, k2, desc);
+}) : (function(o, m, k, k2) {
+    if (k2 === undefined) k2 = k;
+    o[k2] = m[k];
+}));
+var __setModuleDefault = (this && this.__setModuleDefault) || (Object.create ? (function(o, v) {
+    Object.defineProperty(o, "default", { enumerable: true, value: v });
+}) : function(o, v) {
+    o["default"] = v;
+});
+var __importStar = (this && this.__importStar) || function (mod) {
+    if (mod && mod.__esModule) return mod;
+    var result = {};
+    if (mod != null) for (var k in mod) if (k !== "default" && Object.prototype.hasOwnProperty.call(mod, k)) __createBinding(result, mod, k);
+    __setModuleDefault(result, mod);
+    return result;
+};
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
     function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
     return new (P || (P = Promise))(function (resolve, reject) {
@@ -40,17 +63,28 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.Earnings = void 0;
-var custom_http_ts_1 = require("../services/custom-http.ts");
+var custom_http_1 = require("../services/custom-http");
 var config_1 = __importDefault(require("../../config/config"));
-var show_categories_ts_1 = require("../services/show-categories.ts");
+var show_categories_1 = require("../services/show-categories");
 var sidebar_1 = require("./sidebar");
+var bootstrap = __importStar(require("bootstrap"));
 var Earnings = /** @class */ (function () {
     function Earnings() {
         this.editCategoryButtons = null;
         this.deleteCategoryButtons = null;
+        this.addCategoryButton = document.getElementById('add-category');
+        if (this.addCategoryButton) {
+            this.addCategoryButton.onclick = function () {
+                location.href = "#/add-earnings";
+            };
+        }
         //определяем параметры модальных окон
-        this.resultModal = new bootstrap.Modal(document.getElementById('textModal'));
-        this.confirmationModal = new bootstrap.Modal(document.getElementById('confirmationModal'));
+        var textModalElement = document.getElementById('textModal');
+        var confirmationModalElement = document.getElementById('confirmationModal');
+        if (textModalElement && confirmationModalElement) {
+            this.resultModal = new bootstrap.Modal(textModalElement);
+            this.confirmationModal = new bootstrap.Modal(confirmationModalElement);
+        }
         this.modalMessageField = document.getElementById('textModal-message');
         this.textMessage = null;
         this.getCategories();
@@ -66,16 +100,16 @@ var Earnings = /** @class */ (function () {
                         _a.label = 2;
                     case 2:
                         _a.trys.push([2, 9, , 10]);
-                        return [4 /*yield*/, custom_http_ts_1.CustomHttp.request(config_1.default.host + '/categories/income', 'GET')];
+                        return [4 /*yield*/, custom_http_1.CustomHttp.request(config_1.default.host + '/categories/income', 'GET')];
                     case 3:
                         result = _a.sent();
                         if (!result) return [3 /*break*/, 8];
                         if (!(result.error || !result)) return [3 /*break*/, 5];
-                        return [4 /*yield*/, this.showResult(result.message)];
+                        return [4 /*yield*/, this.showResult(result)];
                     case 4:
                         _a.sent();
                         throw new Error(result.message);
-                    case 5: return [4 /*yield*/, show_categories_ts_1.ShowCategories.init(result)];
+                    case 5: return [4 /*yield*/, show_categories_1.ShowCategories.init(result)];
                     case 6:
                         _a.sent(); //отрисуем карточки категорий
                         return [4 /*yield*/, this.processCategories()];
@@ -95,15 +129,11 @@ var Earnings = /** @class */ (function () {
     ;
     Earnings.prototype.processCategories = function () {
         return __awaiter(this, void 0, void 0, function () {
-            var _loop_1, _i, _a, element;
+            var deleteButtonsArray, _loop_1, _i, deleteButtonsArray_1, element;
             var _this = this;
-            return __generator(this, function (_b) {
+            return __generator(this, function (_a) {
                 this.editCategoryButtons = document.querySelectorAll('[id^="edit-"]');
                 this.deleteCategoryButtons = document.querySelectorAll('[id^="delete-"]');
-                this.addCategoryButton = document.getElementById('add-category');
-                this.addCategoryButton.onclick = function () {
-                    location.href = "#/add-earnings";
-                };
                 this.editCategoryButtons.forEach(function (element) {
                     element.addEventListener("click", function () {
                         var id = element.id;
@@ -111,6 +141,7 @@ var Earnings = /** @class */ (function () {
                         location.href = '#/edit-earnings?=' + number;
                     });
                 });
+                deleteButtonsArray = Array.from(this.deleteCategoryButtons);
                 _loop_1 = function (element) {
                     element.addEventListener("click", function () { return __awaiter(_this, void 0, void 0, function () {
                         var id, number;
@@ -130,8 +161,8 @@ var Earnings = /** @class */ (function () {
                         });
                     }); });
                 };
-                for (_i = 0, _a = this.deleteCategoryButtons; _i < _a.length; _i++) {
-                    element = _a[_i];
+                for (_i = 0, deleteButtonsArray_1 = deleteButtonsArray; _i < deleteButtonsArray_1.length; _i++) {
+                    element = deleteButtonsArray_1[_i];
                     _loop_1(element);
                 }
                 return [2 /*return*/];
@@ -146,25 +177,31 @@ var Earnings = /** @class */ (function () {
                         var deleteButton = document.getElementById('delete');
                         var cancelButton = document.getElementById('cancel');
                         var confirmationText = document.getElementById('confirmationText');
-                        confirmationText.innerText = 'Вы действительно хотите удалить категорию?';
+                        if (confirmationText) {
+                            confirmationText.innerText = 'Вы действительно хотите удалить категорию?';
+                        }
                         _this.confirmationModal.show();
-                        cancelButton.onclick = function () {
-                            _this.confirmationModal.hide();
-                            resolve(); // Разрешаем обещание после закрытия модального окна
-                        };
-                        deleteButton.onclick = function () { return __awaiter(_this, void 0, void 0, function () {
-                            return __generator(this, function (_a) {
-                                switch (_a.label) {
-                                    case 0:
-                                        this.confirmationModal.hide();
-                                        return [4 /*yield*/, this.deleteCategory(categoryId)];
-                                    case 1:
-                                        _a.sent();
-                                        resolve(); // Разрешаем обещание после удаления категории
-                                        return [2 /*return*/];
-                                }
-                            });
-                        }); };
+                        if (cancelButton) {
+                            cancelButton.onclick = function () {
+                                _this.confirmationModal.hide();
+                                resolve(); // Разрешаем обещание после закрытия модального окна
+                            };
+                        }
+                        if (deleteButton) {
+                            deleteButton.onclick = function () { return __awaiter(_this, void 0, void 0, function () {
+                                return __generator(this, function (_a) {
+                                    switch (_a.label) {
+                                        case 0:
+                                            this.confirmationModal.hide();
+                                            return [4 /*yield*/, this.deleteCategory(categoryId)];
+                                        case 1:
+                                            _a.sent();
+                                            resolve(); // Разрешаем обещание после удаления категории
+                                            return [2 /*return*/];
+                                    }
+                                });
+                            }); };
+                        }
                     })];
             });
         });
@@ -179,12 +216,12 @@ var Earnings = /** @class */ (function () {
                         _a.label = 1;
                     case 1:
                         _a.trys.push([1, 7, , 8]);
-                        return [4 /*yield*/, custom_http_ts_1.CustomHttp.request(config_1.default.host + '/categories/income/' + categoryId, 'DELETE')];
+                        return [4 /*yield*/, custom_http_1.CustomHttp.request(config_1.default.host + '/categories/income/' + categoryId, 'DELETE')];
                     case 2:
                         result = _a.sent();
                         if (!result) return [3 /*break*/, 6];
                         if (!(result.error || !result)) return [3 /*break*/, 4];
-                        return [4 /*yield*/, this.showResult(result.message)];
+                        return [4 /*yield*/, this.showResult(result)];
                     case 3:
                         _a.sent();
                         throw new Error(result.message);
@@ -207,12 +244,15 @@ var Earnings = /** @class */ (function () {
             var _this = this;
             return __generator(this, function (_a) {
                 return [2 /*return*/, new Promise(function (resolve) {
+                        var _a;
                         _this.textMessage = message.error ? message.message :
-                            "Категория успешно удалена." + "\nСообщение сервера: " + JSON.stringify(message);
-                        _this.modalMessageField.innerText = _this.textMessage;
+                            "Категория успешно удалена." + "\nСообщение сервера: " + JSON.stringify(message.message);
+                        if (_this.modalMessageField) {
+                            _this.modalMessageField.innerText = _this.textMessage;
+                        }
                         _this.resultModal.show();
                         // Обработчик события при закрытии попапа
-                        _this.resultModal._element.addEventListener('hidden.bs.modal', function () {
+                        (_a = _this.resultModal.getElement()) === null || _a === void 0 ? void 0 : _a.addEventListener('hidden.bs.modal', function () {
                             resolve(); // Разрешаем обещание при закрытии попапа
                         });
                     })];
